Day 4 Slide 11 Instruction Decode
In this lab the function fields, registers and opcode are extracted from the instruction. Validity has also been added for the function and register decoding since not all instructions contain these fields. Validity was added mainly to help with debugging since it blanks out values in the waveform when they are invalid making it easy to check the waves visually.