Upload Day 2 Lab Pictures
On day 2 the focus was on the application binary interface (ABI) and basic verification flow. 

The first task was to write an assembly program to be called by a C program and ensure that the result is as expected. This was achieved by using the RISC-V ABI which specifies how to use the registers such as for passing arguments to functions and reading the return values.

The second task was to clone the PicoRV32 git repository and run a C program on the processor. This was achieved by running the script provided which compiles the test C program, converts it into a hexadecimal file and runs the testbench which reads the hexadecimal file for the processor. The expected output was printed on the console as a result of the simulation.